Blacklist, multiple entries and better doc
Implement the use of a permanent blacklist of bad endpoint IPs.

Add option to add multiple whitelisted IPs to the hosts file or
just the best one.

Improve documentation of the script. There is a section for
installation and usage at the top.

Other minor changes have to do with cleanup and organization.